megakernel backward: fp32 master weights, + a grader case that catches it
`dl @ rw.t()` assumed rw shared the activation dtype. True in the bench, false in a real
model, which keeps fp32 master weights and narrows via autocast -- so the forward passed
and the first loss.backward() of a training run raised
"expected mat1 and mat2 to have the same dtype".

Contract in the model's layout: match dtypes for the matmul, and hand autograd a gradient
in the PARAMETER's dtype, not the activation's. The contraction stays in the activation
dtype (what eager does under autocast, and the arrangement the fp64 grader measured at
55-68x closer than eager), so accuracy is unchanged.

grade_fp32_master() reruns the backward in the real layout and asserts gradients exist,
match parameter dtypes and are finite. The old grader ran bf16 weights only, so it could
not see this class of bug.

+++ kernels/sm120/megakernel/moe/block.py
@@ -54,8 +54,15 @@ class _NormRouter(torch.autograd.Function):
         d_logits.scatter_(1, idxl, d_logit_sel)
 
         dl = d_logits.to(x.dtype)
-        d_hn = g_hn + dl @ rw.t()                             # router's contribution to d hn
-        d_rw = hn.t() @ dl
+        # rw is bf16 in the bench but fp32 in a real model, which keeps master weights in fp32 and
+        # relies on autocast to narrow them. Both matmul operands must agree, and the gradient
+        # handed back to autograd must match the PARAMETER's dtype, not the activation's -- writing
+        # this against the bench's dtype is what made the forward pass and the backward die.
+        # Contracting in the activation dtype (rather than promoting to fp32) is deliberate: it is
+        # what eager does under autocast, and it is the arrangement the fp64 grader scored at
+        # 55-68x closer than eager, so matching it preserves the measured accuracy.
+        d_hn = g_hn + dl @ rw.to(dl.dtype).t()                # router's contribution to d hn
+        d_rw = (hn.t() @ dl).to(rw.dtype)
 
         # ---- rmsnorm backward, FUSED. The PyTorch form built five [T,H] fp32 temporaries
         # (~670 MB of traffic at T=65536) against liger's single 0.280 ms kernel -- that was the
